requirements.txt
yt-dlp
opencv-python
moviepy
openai-whisper
googletrans==3.1.0a0
pillow
numpy
requests
torch
transformers 